% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_vcf.R
\name{clean_vcf}
\alias{clean_vcf}
\title{Title Read vcf file and translate into clean vcf}
\usage{
clean_vcf(vcf, filter = c(FALSE, "EMS"))
}
\arguments{
\item{vcf}{Input a vcf dataframe or a vcf file path}

\item{filter}{FALSE is not to filter vcf file, "EMS" is to select G->A and C->T}
}
\value{
retun a clean vcf dataframe
}
\description{
Title Read vcf file and translate into clean vcf
}
