% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onestep_mutmap.R
\name{onestep_mutmap}
\alias{onestep_mutmap}
\title{Title Onestep analysis mutmap based on vcf file and fixed windows}
\usage{
onestep_mutmap(vcf, windows)
}
\arguments{
\item{vcf}{Input a vcf file path}

\item{windows}{Set fixed Windows in Chromosome (bp)}
}
\value{
A plot out SNP-index and ED based on Mutmap
}
\description{
Title Onestep analysis mutmap based on vcf file and fixed windows
}
