% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mutmap.R
\name{plot_mutmap}
\alias{plot_mutmap}
\title{Title Plot SNP-index in Mutmap}
\usage{
plot_mutmap(windows_state, ED = c(FALSE, TRUE))
}
\arguments{
\item{windows_state}{Input a dataframe with snpindex based on windows}

\item{ED}{Whether the windows_state contain ED value}
}
\value{
a Plot with SNP-index
}
\description{
Title Plot SNP-index in Mutmap
}
