% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf.R
\name{read_vcf}
\alias{read_vcf}
\title{Title Read vcf file in R}
\usage{
read_vcf(vcf, header = TRUE, sep = "\\t", comment = "##")
}
\arguments{
\item{vcf}{vcf file is a data frame from GATK output}

\item{header}{The header of vcf file, the default value is TRUE}

\item{sep}{The delimter character of vcf file, the default value is tab}

\item{comment}{The comment character of vcf file, the default value is "##"}
}
\value{
The output is the raw vcf file
}
\description{
Title Read vcf file in R
}
\examples{
\dontrun{vcf <- read_vcf(vcf = "./all.filter.snp.vcf")}
}
