% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows_state.R
\name{windows_state}
\alias{windows_state}
\title{Title Compute SNP-index and ED value in based on fixed Windows in Chromosome}
\usage{
windows_state(vcf_snpindex, ED = c(FALSE, TRUE), windows)
}
\arguments{
\item{vcf_snpindex}{Input a vcf object (dataframe)}

\item{ED}{Whether to calculate ED}

\item{windows}{Set fixed Windows in Chromosome (bp)}
}
\value{
return a dataframe per fixed windows in Chromosome
}
\description{
Title Compute SNP-index and ED value in based on fixed Windows in Chromosome
}
